+++ blueprints/flexberry-list-form/index.js
@@ -68,7 +68,11 @@ module.exports = {
             entityName: options.entity.name,
             modelName: listFormBlueprint.listForm.projections[0].modelName,
             modelProjection: listFormBlueprint.listForm.projections[0].modelProjection,
-            caption: listFormBlueprint.listForm.caption // for use in files\__root__\templates\__name__.hbs
+            caption: listFormBlueprint.listForm.caption, // for use in files\__root__\templates\__name__.hbs
+            importFormRouteName: listFormBlueprint.importFormRouteName,
+            importFormRoutePath: listFormBlueprint.importFormRoutePath,
+            importFormControllerName: listFormBlueprint.importFormControllerName,
+            importFormControllerPath: listFormBlueprint.importFormControllerPath,
         }, listFormBlueprint.locales.getLodashVariablesProperties() // for use in files\__root__\locales\**\forms\__name__.js
         );
     }
@@ -76,15 +80,54 @@ module.exports = {
 var ListFormBlueprint = /** @class */ (function () {
     function ListFormBlueprint(blueprint, options) {
         var listFormsDir = path.join(options.metadataDir, "list-forms");
+        const entityName = options.entity.name;
+        
         if (!options.file) {
-            options.file = options.entity.name + ".json";
+            options.file = entityName + ".json";
         }
-        var localePathTemplate = this.getLocalePathTemplate(options, blueprint.isDummy, path.join("forms", options.entity.name + ".js"));
-        this.locales = new Locales_1.default(options.entity.name, "ru", localePathTemplate);
+
+        var localePathTemplate = this.getLocalePathTemplate(options, blueprint.isDummy, path.join("forms", entityName + ".js"));
+        this.locales = new Locales_1.default(entityName, "ru", localePathTemplate);
         var listFormFile = path.join(listFormsDir, options.file);
         var content = stripBom(fs.readFileSync(listFormFile, "utf8"));
         this.listForm = JSON.parse(content);
         this.locales.setupForm(this.listForm);
+
+        var configsFile = path.join('vendor/flexberry/custom-generator-options/generator-options.json');
+        const defaultImportFormRouteName = 'ListFormRoute';
+        const defaultImportFormRoutePath = 'ember-flexberry/routes/list-form';
+        const defaultImportFormControllerName = 'ListFormController';
+        const defaultImportFormControllerPath = 'ember-flexberry/controllers/list-form';
+        
+        if (fs.existsSync(configsFile)) {
+            var configs = JSON.parse(stripBom(fs.readFileSync(configsFile, "utf8")));
+            
+            if (configs.listForms === undefined) {
+                this.importFormRouteName = defaultImportFormRouteName;
+                this.importFormRoutePath = defaultImportFormRoutePath;
+                this.importFormControllerName = defaultImportFormControllerName;
+                this.importFormControllerPath = defaultImportFormControllerPath;
+            } else {
+                const { baseRoute, baseController } = configs.listForms.defaultForm;
+
+                if (configs.listForms[entityName] !== undefined) {
+                    this.importFormRouteName = configs.listForms[entityName].baseRoute.name;
+                    this.importFormRoutePath = configs.listForms[entityName].baseRoute.path;
+                    this.importFormControllerName = configs.listForms[entityName].baseController.name;
+                    this.importFormControllerPath = configs.listForms[entityName].baseController.path;
+                } else if (configs.listForms.defaultForm !== undefined) {
+                    this.importFormRouteName = baseRoute.name;
+                    this.importFormRoutePath = baseRoute.path;
+                    this.importFormControllerName = baseController.name;
+                    this.importFormControllerPath = baseController.path;
+                };
+            };
+        } else {
+            this.importFormRouteName = defaultImportFormRouteName;
+            this.importFormRoutePath = defaultImportFormRoutePath;
+            this.importFormControllerName = defaultImportFormControllerName;
+            this.importFormControllerPath = defaultImportFormControllerPath;
+        };
     }
     ListFormBlueprint.prototype.getLocalePathTemplate = function (options, isDummy, localePathSuffix) {
         var targetRoot = "app";